ml-service/requirements.txt
fastapi
uvicorn
pymysql
pandas
scikit-learn
sqlalchemy
python-dotenv